Revert "Adding Dashboard Widget & Report Plugin Functionality"
