--- report-plugin/src/main/resources/META-INF/resources/WEB-INF/views/applications/widgets/reportPlugin.jsp
@@ -1,3 +1,0 @@
-<%@ include file="/common/taglibs.jsp"%>
-
-<%--All jsp code will go here--%>
