--- report-plugin/src/main/resources/META-INF/resources/scripts/report-plugin.js
@@ -1,4 +1,0 @@
-var threadfixModule = angular.module('threadfix');
-
-
-// All js goes here (angular controllers, directives, factories, etc)
