--- report-plugin/src/main/resources/META-INF/web-fragment.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-fragment xmlns="http://java.sun.com/xml/ns/javaee"
-              xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-              xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-	http://java.sun.com/xml/ns/javaee/web-fragment_3_0.xsd"
-              version="3.0">
-
-    <name id="rewrite">pluginShortName</name>
-    <distributable />
-
-    <ordering>
-        <before>
-            <others />
-        </before>
-    </ordering>
-
-
-</web-fragment>
