+++ threadfix-entities/src/main/java/com/denimgroup/threadfix/data/entities/GRCControl.java
@@ -75,6 +75,7 @@ public class GRCControl extends AuditableEntity {
 
     @OneToOne
     @JoinColumn(name = "vulnerabilityId")
+    @JsonView({AllViews.TableRow.class})
     public Vulnerability getVulnerability() {
         return vulnerability;
     }
