+++ threadfix-importers/src/main/java/com/denimgroup/threadfix/importer/loader/AnnotationLoader.java
@@ -128,6 +128,8 @@ public final class AnnotationLoader<T extends java.lang.annotation.Annotation> {
             }
         }
 
+        assert !savedMap.isEmpty() : "Map was empty after loading annotations from classpath.";
+
         return savedMap;
     }
