+++ threadfix-main/src/main/java/com/denimgroup/threadfix/webapp/controller/SystemSettingsController.java
@@ -24,11 +24,9 @@
 
 package com.denimgroup.threadfix.webapp.controller;
 
-import com.denimgroup.threadfix.data.entities.DashboardWidget;
 import com.denimgroup.threadfix.data.entities.DefaultConfiguration;
 import com.denimgroup.threadfix.data.entities.Role;
 import com.denimgroup.threadfix.logging.SanitizedLogger;
-import com.denimgroup.threadfix.service.DashboardWidgetService;
 import com.denimgroup.threadfix.service.DefaultConfigService;
 import com.denimgroup.threadfix.service.RoleService;
 import com.denimgroup.threadfix.service.enterprise.EnterpriseTest;
@@ -56,8 +54,6 @@ public class SystemSettingsController {
 	private RoleService roleService = null;
 	@Autowired
     private DefaultConfigService defaultConfigService = null;
-	@Autowired
-	private DashboardWidgetService dashboardWidgetService = null;
 	
 	@InitBinder
 	public void setAllowedFields(WebDataBinder dataBinder) {
@@ -67,8 +63,7 @@ public class SystemSettingsController {
                     "proxyHost", "proxyPort", "proxyUsername", "proxyPassword", "shouldProxyVeracode",
                     "shouldProxyQualys", "shouldProxyTFS", "shouldProxyBugzilla", "shouldProxyJira",
                     "shouldProxyVersionOne", "shouldProxyHPQC", "shouldProxyWhiteHat", "shouldProxyTrustwaveHailstorm",
-					"shouldProxyContrast", "shouldUseProxyCredentials", "sessionTimeout", "dashboardTopLeftId", 
-                    "dashboardTopRightId", "dashboardBottomLeftId", "dashboardBottomRightId");
+					"shouldProxyContrast", "shouldUseProxyCredentials", "sessionTimeout");
 		} else {
             // this should prevent any parameters from coming in.
             // We also need to check permissions on the server side though
@@ -79,11 +74,6 @@ public class SystemSettingsController {
 	@ModelAttribute
 	public List<Role> populateRoles() {
 		return roleService.loadAll();
-	}
-
-	@ModelAttribute("dashboardWidgets")
-	public List<DashboardWidget> populateDashboardWidgetTypes() {
-		return dashboardWidgetService.loadAllAvailable();
 	}
 
 	@RequestMapping(method = RequestMethod.GET)
