+++ threadfix-main/src/main/java/com/denimgroup/threadfix/webapp/filter/CacheBustFilter.java
@@ -53,7 +53,7 @@ public class CacheBustFilter extends GenericFilterBean {
     private final SanitizedLogger log = new SanitizedLogger(CacheBustFilter.class);
     private String gitCommit = null;
     private String buildNumber = null;
-    private Date buildDate = Calendar.getInstance().getTime();
+    private Date buildDate = null;
 
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
@@ -62,7 +62,7 @@ public class CacheBustFilter extends GenericFilterBean {
         // If there was a build number defined in the war, then use it for the cache buster.
         req.setAttribute("gitCommit", gitCommit);
         req.setAttribute("buildNumber", (buildNumber != null) ? buildNumber : "2.2-SNAPSHOT" + "-" + gitCommit);
-        req.setAttribute("buildDate", buildDate);
+        req.setAttribute("buildDate", (buildDate != null) ? buildDate : Calendar.getInstance().getTime());
 
         chain.doFilter(request, response);
     }
@@ -82,6 +82,13 @@ public class CacheBustFilter extends GenericFilterBean {
                 String date = attrs.getValue("Implementation-Build-Date");
                 gitCommit = attrs.getValue("Implementation-Build");
 
+                gitCommit = (gitCommit != null) ? gitCommit : String.valueOf(new Random().nextInt(10000000));
+
+                if (version != null && gitCommit != null){
+                    buildNumber = version + "-" + gitCommit;
+                    log.info("Application version set to: " + buildNumber);
+                }
+
                 if(date != null) {
                     SimpleDateFormat dt = new SimpleDateFormat("yyyy-MM-dd hh:mm:ss");
                     try {
@@ -91,14 +98,6 @@ public class CacheBustFilter extends GenericFilterBean {
                     }
                 }
 
-                // build fake git commit # for dev env
-                SimpleDateFormat dt = new SimpleDateFormat("yyyyMMddhh");
-                gitCommit = (gitCommit != null) ? gitCommit : dt.format(buildDate);
-
-                if (version != null){
-                    buildNumber = version + "-" + gitCommit;
-                    log.info("Application version set to: " + buildNumber);
-                }
             }
         } catch (IOException e) {
             log.error("I/O Exception reading manifest: " + e.getMessage());
