+++ threadfix-main/src/main/webapp/WEB-INF/views/applications/detail.jsp
@@ -10,8 +10,7 @@
 	<cbs:cachebustscript src="/scripts/grc-control-submission-modal-controller.js"/>
 	<cbs:cachebustscript src="/scripts/modal-controller-with-config.js"/>
     <cbs:cachebustscript src="/scripts/edit-application-modal-controller.js"/>
-	<cbs:cachebustscript src="/scripts/left-report-controller.js"/>
-	<cbs:cachebustscript src="/scripts/right-report-controller.js"/>
+	<cbs:cachebustscript src="/scripts/reports-controller.js"/>
 	<cbs:cachebustscript src="/scripts/scan-table-controller.js"/>
 	<cbs:cachebustscript src="/scripts/upload-scan-controller.js"/>
 	<cbs:cachebustscript src="/scripts/scheduled-scan-tab-controller.js"/>
@@ -48,11 +47,7 @@
         </div>
 
         <div class="container-fluid">
-            <div class="row-fluid">
-                <c:set var="csrfToken" value="${ emptyUrl }" scope="request"/>
-                <jsp:include page="/WEB-INF/views/applications/widgets/vulnerabilityTrending.jsp"/>
-                <jsp:include page="/WEB-INF/views/applications/widgets/mostVulnerableApps.jsp"/>
-            </div>
+            <%@include file="reports.jspf"%>
         </div>
 
         <tabset style="margin-top:10px;">
