+++ threadfix-main/src/main/webapp/WEB-INF/views/applications/reports.jspf
@@ -1,0 +1,59 @@
+<div ng-controller="ReportsController" ng-init="csrfToken = '<c:out value="${ emptyUrl }"/>'">
+    <c:if test="${ canGenerateReports }">
+        <div class="row-fluid">
+            <div class="span6">
+                <h4>Vulnerability Trending<span style="font-size:12px;float:right;">
+                    <a id="leftViewMore" ng-href="{{ urlRoot }}/reports/9{{ seeMoreExtension }}{{ csrfToken }}">View More</a></span>
+                </h4>
+                <div id="leftTileReport">
+                    <div ng-show="trendingScansData" class="team-report-wrapper report-image">
+                        <d3-trending data="trendingScansData" width="422" height="250" margin="leftMargin"
+                                     start-date="trendingStartDate" end-date="trendingEndDate"></d3-trending>
+                    </div>
+                    <div ng-hide="noData || trendingScansData || leftReportFailed" class="team-report-wrapper report-image">
+                        <div style="float:right;padding-top:120px" class="modal-loading"><div><span class="spinner dark"></span>Loading...</div></div>
+                    </div>
+                    <div ng-show="leftReportFailed" class="team-report-wrapper report-image">
+                        <div class="text">
+                            Report Failed
+                        </div>
+                    </div>
+                    <div ng-show="noData && !allScans" class="team-report-wrapper report-image">
+                        <div class="text">
+                            No Data Found
+                        </div>
+                    </div>
+                    <div ng-show="noData && allScans" class="team-report-wrapper report-image">
+                        <div class="text">
+                            No New Scans Uploaded This Time
+                        </div>
+                    </div>
+                </div>
+            </div>
+
+            <div class="span6">
+                <h4>{{ rightReportTitle }}<span style="font-size:12px;float:right;">
+                    <a id="rightViewMore" ng-href="{{ urlRoot }}/reports/10{{ seeMoreExtension }}{{ csrfToken }}">View More</a></span>
+                </h4>
+                <div id="rightTileReport">
+                    <div ng-show="topAppsData" class="team-report-wrapper report-image">
+                        <d3-hbars data="topAppsData" label = "label" width="422" height="250" margin="rightMargin"></d3-hbars>
+                    </div>
+                    <div ng-hide="empty || topAppsData || rightReportFailed" class="team-report-wrapper report-image">
+                        <div style="float:right;padding-top:120px" class="modal-loading"><div><span class="spinner dark"></span>Loading...</div></div>
+                    </div>
+                    <div ng-show="rightReportFailed" class="team-report-wrapper report-image">
+                        <div class="text">
+                            Report Failed
+                        </div>
+                    </div>
+                    <div ng-show="empty && !topAppsData" class="team-report-wrapper report-image">
+                        <div class="text">
+                            No Data Found
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </c:if>
+</div>
