--- threadfix-main/src/main/webapp/WEB-INF/views/applications/widgets/vulnerabilityTrending.jsp
@@ -1,33 +1,0 @@
-<%@ include file="/common/taglibs.jsp"%>
-
-<div class="span6" ng-controller="LeftReportController" ng-init="csrfToken = '${ csrfToken }'">
-  <c:if test="${ canGenerateReports }">
-    <h4>Vulnerability Trending<span style="font-size:12px;float:right;">
-      <a id="leftViewMore" ng-href="{{ urlRoot }}/reports/9{{ seeMoreExtension }}{{ csrfToken }}">View More</a></span>
-    </h4>
-    <div id="leftTileReport">
-      <div ng-show="trendingScansData" class="team-report-wrapper report-image">
-        <d3-trending data="trendingScansData" width="422" height="250" margin="leftMargin"
-             start-date="trendingStartDate" end-date="trendingEndDate"></d3-trending>
-      </div>
-      <div ng-hide="noData || trendingScansData || leftReportFailed" class="team-report-wrapper report-image">
-        <div style="float:right;padding-top:120px" class="modal-loading"><div><span class="spinner dark"></span>Loading...</div></div>
-      </div>
-      <div ng-show="leftReportFailed" class="team-report-wrapper report-image">
-        <div class="text">
-          Report Failed
-        </div>
-      </div>
-      <div ng-show="noData && !allScans" class="team-report-wrapper report-image">
-        <div class="text">
-          No Data Found
-        </div>
-      </div>
-      <div ng-show="noData && allScans" class="team-report-wrapper report-image">
-        <div class="text">
-          No New Scans Uploaded This Time
-        </div>
-      </div>
-    </div>
-  </c:if>
-</div>
