+++ threadfix-main/src/main/webapp/WEB-INF/web.xml
@@ -234,8 +234,6 @@
                 protected-regex ^/configuration/scanqueue/scanQueueTask/[0-9]+/delete$,
         		protected-regex ^/configuration/scanqueue/organizations/[0-9]+/applications/[0-9]+/scanQueueTask/[0-9]+/delete$,
         		protected-regex ^/configuration/scanqueue/organizations/[0-9]+/applications/[0-9]+/addScanQueueTask$,
-                protected-regex ^/graphConfig$, protected-regex ^/graphConfig/table$, protected-regex ^/graphConfig/channels$,
-                protected-regex ^/graphConfig/data,
         	</param-value>
 		</init-param>
 	</filter>
