+++ threadfix-main/src/main/webapp/scripts/report/directives/d3-dashboards.js
@@ -91,6 +91,7 @@ d3ThreadfixModule.directive('d3Vbars', ['$window', '$timeout', 'd3', 'd3Service'
                             return colDuration*drawTime; }) ;
 
                 };
+                ;
             }
         }
     }]);
