--- threadfix-service-interfaces/src/main/java/com/denimgroup/threadfix/service/DashboardWidgetService.java
@@ -1,92 +1,0 @@
-////////////////////////////////////////////////////////////////////////
-//
-//     Copyright (c) 2009-2015 Denim Group, Ltd.
-//
-//     The contents of this file are subject to the Mozilla Public License
-//     Version 2.0 (the "License"); you may not use this file except in
-//     compliance with the License. You may obtain a copy of the License at
-//     http://www.mozilla.org/MPL/
-//
-//     Software distributed under the License is distributed on an "AS IS"
-//     basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
-//     License for the specific language governing rights and limitations
-//     under the License.
-//
-//     The Original Code is ThreadFix.
-//
-//     The Initial Developer of the Original Code is Denim Group, Ltd.
-//     Portions created by Denim Group, Ltd. are Copyright (C)
-//     Denim Group, Ltd. All Rights Reserved.
-//
-//     Contributor(s): Denim Group, Ltd.
-//
-////////////////////////////////////////////////////////////////////////
-package com.denimgroup.threadfix.service;
-
-import com.denimgroup.threadfix.data.entities.DashboardWidget;
-
-import java.util.List;
-
-/**
- * @author zabdisubhan
- *
- */
-public interface DashboardWidgetService {
-
-    boolean isInitialized();
-
-    void setInitialized(boolean initialized);
-
-    /**
-     * @return List<DashboardWidget>
-     */
-    List<DashboardWidget> loadAll();
-
-   /**
-     * @return List<DashboardWidget>
-     */
-    List<DashboardWidget> loadAllAvailable();
-
-    /**
-     * @return List<DashboardWidget>
-     */
-    List<DashboardWidget> loadByIds(List<Integer> dashboardWidgetIds);
-
-    /**
-     * @return List<DashboardWidget>
-     */
-    List<DashboardWidget> loadAllNativeReports();
-
-    /**
-     * @return List<DashboardWidget>
-     */
-    List<DashboardWidget> loadAllNonNativeReports();
-
-    /**
-     * @param dashboardWidgetId
-     * @return DashboardWidget
-     */
-    DashboardWidget load(int dashboardWidgetId);
-
-    /**
-     * @param widgetName
-     * @return DashboardWidget
-     */
-    DashboardWidget load(String widgetName);
-
-    /**
-     * @param dashboardWidget
-     */
-    void store(DashboardWidget dashboardWidget);
-
-    /**
-     * @param dashboardWidget
-     */
-    void delete(DashboardWidget dashboardWidget);
-
-    /**
-     * @param dashboardWidgetId
-     */
-    void deleteById(int dashboardWidgetId);
-
-}
